Implementation/FleetManager/requirements.txt
connexion >=2.6.0,<3.0.0
connexion[swagger-ui] >=2.6.0,<3.0.0
python_dateutil ==2.6.0
setuptools >=21.0.0
swagger-ui-bundle >=0.0.2
grpcio
grpcio-tools
concurrent-log-handler
numpy
catkin_pkg
empy
lark
mariadb
paho.mqtt
flask 
requests